+++ app/Http/Controllers/RequestController.php
@@ -12,7 +12,6 @@ use App\Utilities\Response as UtilityResponse;
 use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Config;
 
 /**
  * Manage categories
@@ -61,13 +60,13 @@ class RequestController extends Controller
     }
 
     /**
-     * Return the paginated request log
+     * Return the paginated access log
      *
      * @param Request $request
      *
      * @return JsonResponse
      */
-    public function log(Request $request): JsonResponse
+    public function accessLog(Request $request): JsonResponse
     {
         $total = (new RequestLog())->totalCount();
 
@@ -98,41 +97,19 @@ class RequestController extends Controller
     }
 
     /**
-     * Return a summary of the monthly requests
-     *
-     * @param Request $request
-     *
-     * @return JsonResponse
-     */
-    public function monthlyRequests(Request $request): JsonResponse
-    {
-        $monthly_summary = (new RequestLog())->monthlyRequests();
-
-        $summary = [];
-        foreach ($monthly_summary as $month) {
-            $summary[$month['year']][] = ['month' => $month['month'], 'requests' => $month['requests']];
-        }
-
-        return response()->json(
-            $summary,
-            200
-        );
-    }
-
-    /**
      * Generate the OPTIONS request for log
      *
      * @param Request $request
      */
-    public function optionsLog(Request $request)
+    public function optionsAccessLog(Request $request)
     {
-        $this->optionsResponse(
+        return $this->generateOptionsForIndex(
             [
-                'GET' => [
-                    'description' => trans('route-descriptions.request_GET_log'),
-                    'authenticated' => false,
-                    'parameters' => []
-                ]
+                'description_localisation' => 'route-descriptions.request_GET_access-log',
+                'parameters_config' => [],
+                'conditionals' => [],
+                'pagination' => true,
+                'authenticated' => false
             ]
         );
     }
@@ -149,6 +126,7 @@ class RequestController extends Controller
                 'description_localisation' => 'route-descriptions.request_GET_error_log',
                 'parameters_config' => 'api.request.parameters.collection',
                 'conditionals' => [],
+                'pagination' => false,
                 'authenticated' => false
             ],
             [
@@ -156,24 +134,6 @@ class RequestController extends Controller
                 'fields_config' => 'api.request.fields',
                 'conditionals' => [],
                 'authenticated' => false
-            ]
-        );
-    }
-
-    /**
-     * Generate the OPTIONS request for monthly request summary
-     *
-     * @param Request $request
-     */
-    public function optionsMonthlyRequests(Request $request)
-    {
-        $this->optionsResponse(
-            [
-                'GET' => [
-                    'description' => trans('route-descriptions.request_GET_log_monthly_requests'),
-                    'authenticated' => false,
-                    'parameters' => []
-                ]
             ]
         );
     }