+++ app/Http/Parameters/Get.php
@@ -85,10 +85,40 @@ class Get
                     }
                     break;
 
+                case 'months':
+                    if (array_key_exists($key, self::$parameters) === true) {
+                        if (General::isBooleanValue(self::$parameters[$key]) === false) {
+                            unset(self::$parameters[$key]);
+                        }
+                    }
+                    break;
+
                 case 'resource_type':
                     if (array_key_exists($key, self::$parameters) === true) {
                         if ((new ResourceType())->
                             where('id', '=', self::$parameters[$key])->exists() === false) {
+                            unset(self::$parameters[$key]);
+                        }
+                    }
+                    break;
+
+                case 'resource-type':
+                    if (array_key_exists($key, self::$parameters) === true) {
+                        if ((new ResourceType())->
+                            where('id', '=', self::$parameters[$key])->exists() === false) {
+                            unset(self::$parameters[$key]);
+                        }
+                    }
+                    break;
+
+                case 'subcategory':
+                    if (array_key_exists($key, self::$parameters) === true) {
+                        if (
+                            (new SubCategory())->
+                            where('sub_category.id', '=', self::$parameters[$key])->
+                            where('sub_category.category_id', '=', self::$parameters['category'])->
+                            exists() === false
+                        ) {
                             unset(self::$parameters[$key]);
                         }
                     }
@@ -107,11 +137,27 @@ class Get
                     }
                     break;
 
+                case 'subcategories':
+                    if (array_key_exists($key, self::$parameters) === true) {
+                        if (General::isBooleanValue(self::$parameters[$key]) === false) {
+                            unset(self::$parameters[$key]);
+                        }
+                    }
+                    break;
+
                 case 'year':
                     if (array_key_exists($key, self::$parameters) === true) {
                         if (intval(self::$parameters[$key]) < 2013 ||
                             self::$parameters[$key] > intval(date('Y'))) {
 
+                            unset(self::$parameters[$key]);
+                        }
+                    }
+                    break;
+
+                case 'years':
+                    if (array_key_exists($key, self::$parameters) === true) {
+                        if (General::isBooleanValue(self::$parameters[$key]) === false) {
                             unset(self::$parameters[$key]);
                         }
                     }